+++ modules/features/account/src/main/java/au/com/shiftyjelly/pocketcasts/account/AccountActivity.kt
@@ -1,18 +1,21 @@
 package au.com.shiftyjelly.pocketcasts.account
 
-import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
+import androidx.navigation.NavDestination
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
 import au.com.shiftyjelly.pocketcasts.account.databinding.AccountActivityBinding
+import au.com.shiftyjelly.pocketcasts.account.viewmodel.CreateAccountState
 import au.com.shiftyjelly.pocketcasts.account.viewmodel.CreateAccountViewModel
 import au.com.shiftyjelly.pocketcasts.account.viewmodel.SubscriptionType
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.AnalyticsHelper
 import au.com.shiftyjelly.pocketcasts.utils.Util
@@ -25,7 +28,8 @@ import au.com.shiftyjelly.pocketcasts.images.R as IR
 class AccountActivity : AppCompatActivity() {
 
     @Inject lateinit var theme: Theme
-
+    @Inject lateinit var analyticsTracker: AnalyticsTrackerWrapper
+    private val viewModel: CreateAccountViewModel by viewModels()
     private lateinit var binding: AccountActivityBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -42,7 +46,6 @@ class AccountActivity : AppCompatActivity() {
             val graph = navInflater.inflate(R.navigation.account_nav_graph)
             val arguments = Bundle()
 
-            val viewModel: CreateAccountViewModel by viewModels()
             if (isNewUpgradeInstance(intent)) {
                 viewModel.clearReadyForUpgrade()
                 graph.setStartDestination(R.id.createFrequencyFragment)
@@ -66,9 +69,11 @@ class AccountActivity : AppCompatActivity() {
 
             val navConfiguration = AppBarConfiguration(navController.graph)
             binding.toolbar?.setupWithNavController(navController, navConfiguration)
+            binding.toolbar?.setNavigationOnClickListener { _ -> onBackPressed() }
             binding.carHeader?.btnClose?.setOnClickListener { onBackPressed() }
 
             navController.addOnDestinationChangedListener { _, destination, _ ->
+                destination.trackShown()
                 if (!Util.isCarUiMode(this)) {
                     when (destination.id) {
                         R.id.createDoneFragment, R.id.accountFragment, R.id.promoCodeFragment -> {
@@ -95,6 +100,7 @@ class AccountActivity : AppCompatActivity() {
 
     override fun onBackPressed() {
         val currentFragment = findNavController(R.id.nav_host_fragment).currentDestination
+        currentFragment?.trackDismissed()
         if (currentFragment?.id == R.id.createPayNowFragment || currentFragment?.id == R.id.createDoneFragment) {
             finish()
             return
@@ -105,6 +111,53 @@ class AccountActivity : AppCompatActivity() {
 
         UiUtil.hideKeyboard(binding.root)
         super.onBackPressed()
+    }
+
+    private fun NavDestination.trackShown() {
+        val analyticsEvent = when (id) {
+            R.id.accountFragment -> AnalyticsEvent.SETUP_ACCOUNT_SHOWN
+            R.id.signInFragment -> AnalyticsEvent.SIGNIN_SHOWN
+            R.id.createAccountFragment -> AnalyticsEvent.SELECT_ACCOUNT_TYPE_SHOWN
+            R.id.createEmailFragment -> AnalyticsEvent.CREATE_ACCOUNT_SHOWN
+            R.id.createTOSFragment -> AnalyticsEvent.TERMS_OF_USE_SHOWN
+            R.id.createFrequencyFragment -> AnalyticsEvent.SELECT_PAYMENT_FREQUENCY_SHOWN
+            R.id.createPayNowFragment -> AnalyticsEvent.CONFIRM_PAYMENT_SHOWN
+            R.id.resetPasswordFragment -> AnalyticsEvent.FORGOT_PASSWORD_SHOWN
+            R.id.createDoneFragment -> AnalyticsEvent.ACCOUNT_UPDATED_SHOWN
+            else -> null
+        }
+        val properties = when (id) {
+            R.id.createPayNowFragment -> {
+                val subscription = viewModel.subscription.value
+                subscription?.let { mapOf(PRODUCT_KEY to it.productDetails.productId) }
+            }
+            R.id.createDoneFragment -> {
+                val source = when (viewModel.createAccountState.value) {
+                    CreateAccountState.AccountCreated -> AccountUpdatedSource.CREATE_ACCOUNT.analyticsValue
+                    CreateAccountState.SubscriptionCreated -> AccountUpdatedSource.CONFIRM_PAYMENT.analyticsValue
+                    else -> null
+                }
+                source?.let { mapOf(SOURCE_KEY to source) }
+            }
+            else -> null
+        } ?: emptyMap()
+        analyticsEvent?.let { analyticsTracker.track(it, properties) }
+    }
+
+    private fun NavDestination.trackDismissed() {
+        val analyticsEvent = when (id) {
+            R.id.accountFragment -> AnalyticsEvent.SETUP_ACCOUNT_DISMISSED
+            R.id.signInFragment -> AnalyticsEvent.SIGNIN_DISMISSED
+            R.id.createAccountFragment -> AnalyticsEvent.SELECT_ACCOUNT_TYPE_DISMISSED
+            R.id.createEmailFragment -> AnalyticsEvent.CREATE_ACCOUNT_DISMISSED
+            R.id.createTOSFragment -> AnalyticsEvent.TERMS_OF_USE_DISMISSED
+            R.id.createFrequencyFragment -> AnalyticsEvent.SELECT_PAYMENT_FREQUENCY_DISMISSED
+            R.id.createPayNowFragment -> AnalyticsEvent.CONFIRM_PAYMENT_DISMISSED
+            R.id.resetPasswordFragment -> AnalyticsEvent.FORGOT_PASSWORD_DISMISSED
+            R.id.createDoneFragment -> AnalyticsEvent.ACCOUNT_UPDATED_DISMISSED
+            else -> null
+        }
+        analyticsEvent?.let { analyticsTracker.track(it) }
     }
 
     companion object {
@@ -127,7 +180,8 @@ class AccountActivity : AppCompatActivity() {
         fun isNewAutoSelectPlusInstance(intent: Intent): Boolean {
             return intent.getBooleanExtra(AUTO_SELECT_PLUS, false)
         }
-
+        private const val PRODUCT_KEY = "product"
+        private const val SOURCE_KEY = "source"
         const val IS_PROMO_CODE = "account_activity.is_promo_code"
         const val PROMO_CODE_VALUE = "account_activity.promo_code"
         const val PROMO_CODE_RETURN_DESCRIPTION = "account_activity.promo_code_return_description"
@@ -149,18 +203,12 @@ class AccountActivity : AppCompatActivity() {
         fun isSignInInstance(intent: Intent): Boolean {
             return intent.getBooleanExtra(SIGN_IN_ONLY, false)
         }
+    }
 
-        fun signInInstance(context: Context, intentOnSuccess: PendingIntent?): Intent {
-            val intent = Intent(context, AccountActivity::class.java)
-            intent.putExtra(SIGN_IN_ONLY, true)
-            intent.putExtra(SUCCESS_INTENT, intentOnSuccess)
-            return intent
-        }
-
-        fun supporterInstance(context: Context): Intent {
-            val intent = Intent(context, AccountActivity::class.java)
-            intent.putExtra(SUPPORTER_INTENT, true)
-            return intent
-        }
+    enum class AccountUpdatedSource(val analyticsValue: String) {
+        CREATE_ACCOUNT("create_account"),
+        CONFIRM_PAYMENT("confirm_payment"),
+        CHANGE_EMAIL("change_email"),
+        CHANGE_PASSWORD("change_password"),
     }
 }