+++ modules/services/compose/src/main/java/au/com/shiftyjelly/pocketcasts/compose/Theme.kt
@@ -7,6 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.ui.unit.sp
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 
 private val LocalColors = staticCompositionLocalOf { PocketCastsTheme(colors = ThemeLightColors, isLight = true) }
@@ -61,6 +62,34 @@ private fun SurfacedContent(
 ) {
     Surface(color = MaterialTheme.colors.background) {
         content()
+    }
+}
+
+@Composable
+fun AutomotiveTheme(content: @Composable () -> Unit) {
+    val theme = PocketCastsTheme(colors = ThemeDarkColors, isLight = false)
+    val typography = MaterialTheme.typography
+    // Increase the size of the fonts on Automotive to match the system
+    CompositionLocalProvider(LocalColors provides theme) {
+        MaterialTheme(
+            colors = buildMaterialColors(colors = theme.colors, isLight = theme.isLight),
+            typography = typography.copy(
+                h1 = typography.h1.copy(fontSize = 144.sp),
+                h2 = typography.h2.copy(fontSize = 90.sp),
+                h3 = typography.h3.copy(fontSize = 72.sp),
+                h4 = typography.h4.copy(fontSize = 51.sp),
+                h5 = typography.h5.copy(fontSize = 36.sp),
+                h6 = typography.h6.copy(fontSize = 30.sp),
+                body1 = typography.body1.copy(fontSize = 24.sp),
+                body2 = typography.body2.copy(fontSize = 21.sp),
+                subtitle1 = typography.subtitle1.copy(fontSize = 24.sp),
+                subtitle2 = typography.subtitle2.copy(fontSize = 21.sp),
+                button = typography.button.copy(fontSize = 21.sp),
+                caption = typography.caption.copy(fontSize = 18.sp),
+                overline = typography.overline.copy(fontSize = 15.sp)
+            ),
+            content = content
+        )
     }
 }