+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/type/PodcastsSortType.kt
@@ -13,21 +13,24 @@ enum class PodcastsSortType(
     val serverId: Int,
     val labelId: Int,
     val podcastComparator: Comparator<Podcast>,
-    val folderComparator: Comparator<FolderItem>
+    val folderComparator: Comparator<FolderItem>,
+    val analyticsValue: String,
 ) {
     DATE_ADDED_OLDEST_TO_NEWEST(
         clientId = 0,
         serverId = 0,
         labelId = R.string.podcasts_sort_by_date_added,
         podcastComparator = compareBy { it.addedDate },
-        folderComparator = compareBy { it.addedDate }
+        folderComparator = compareBy { it.addedDate },
+        analyticsValue = "date_added"
     ),
     NAME_A_TO_Z(
         clientId = 2,
         serverId = 1,
         labelId = R.string.name,
         podcastComparator = compareBy { cleanStringForSort(it.title) },
-        folderComparator = compareBy { cleanStringForSort(it.title) }
+        folderComparator = compareBy { cleanStringForSort(it.title) },
+        analyticsValue = "name"
     ),
     EPISODE_DATE_NEWEST_TO_OLDEST(
         clientId = 5,
@@ -35,14 +38,16 @@ enum class PodcastsSortType(
         labelId = R.string.podcasts_sort_by_release_date,
         // use a query to get the podcasts ordered by episode release date
         podcastComparator = Comparator { _, _ -> 0 },
-        folderComparator = Comparator { _, _ -> 0 }
+        folderComparator = Comparator { _, _ -> 0 },
+        analyticsValue = "episode_release_date"
     ),
     DRAG_DROP(
         clientId = 6,
         serverId = 3,
         labelId = R.string.podcasts_sort_by_drag_drop,
         podcastComparator = compareBy { it.sortPosition },
-        folderComparator = compareBy { it.sortPosition }
+        folderComparator = compareBy { it.sortPosition },
+        analyticsValue = "drag_and_drop"
     );
 
     companion object {