+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/FolderManagerImpl.kt
@@ -254,4 +254,6 @@ class FolderManagerImpl @Inject constructor(
             return itemsSorted
         }
     }
+
+    override fun countFolders() = folderDao.count()
 }