+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/PostCallback.kt
@@ -1,6 +1,5 @@
 package au.com.shiftyjelly.pocketcasts.servers
 
-interface PostCallback {
+interface PostCallback : ServerFailure {
     fun onSuccess(data: String?, response: ServerResponse)
-    fun onFailed(errorCode: Int, userMessage: String?, userMessageId: Int?, serverMessage: String?, throwable: Throwable?)
 }