+++ CHANGELOG.md
@@ -1,9 +1,19 @@
+7.23.0
+-----
+
+*   Bug Fixes:
+    *   Fix custom video file not opening in full-screen mode.
+        ([#304](https://github.com/Automattic/pocket-casts-android/pull/304)).
+
 7.22.0
 -----
 
 *   Bug Fixes:
     *   Fix playing on Chromecast always shows buffering.
         ([#254](https://github.com/Automattic/pocket-casts-android/pull/254)).
+    *   Fix Chromecast not taking over playback after connection.
+    *   Fix Plus subscription confirm button not always working.
+        ([#284](https://github.com/Automattic/pocket-casts-android/pull/284)).
 
 7.21.0
 -----